Add possibility to add several keys to the uboot dtb
A new test is added, test-dynamic-manykeys.sh,  where
two keys are added together with the required-mode="any"
property so verification succeeds when one key verifies.

At the same time the possibility to output the data to a text file
is removed, thus not compatible with the upstream ubpubkey.py
any more.